Builds out file actions on the custom file index table.

+++ php/fileshare.php
@@ -101,15 +101,26 @@ function gmuw_fs_index_file_table($posts){
 
 			//do we not have a good value?
 			if ($mypostid==0) {
-				$return_value='<div class="error notice is-dismissable "><p>Bad post ID. Unable to '.$_GET['action'].'.</p></div>';
+				$return_value='<div class="notice notice-error is-dismissable "><p>Bad post ID. Unable to '.$_GET['action'].'.</p></div>';
 			} else {
 
-				if ($_GET['action']=='edit') {
-					$return_value='<div class="updated notice is-dismissable "><p>'.$_GET['action'].' post '.$mypostid.'</p></div>';
-				}
+				//do we not have permissions?
+				if (!(get_post($mypostid)->post_author == get_current_user_id() || current_user_can('manage_options') ) ) {
 
-				if ($_GET['action']=='delete') {
-					$return_value='<div class="updated notice is-dismissable "><p>'.$_GET['action'].' post '.$mypostid.'</p></div>';
+					$return_value='<div class="notice notice-error is-dismissable "><p>You do not have permissions.</p></div>';
+
+				} else {
+
+					//we are good. we have a good id, and we have permissions...
+
+					if ($_GET['action']=='edit') {
+						$return_value='<div class="notice notice-success is-dismissable "><p>'.$_GET['action'].' post '.$mypostid.'</p></div>';
+					}
+
+					if ($_GET['action']=='delete') {
+						$return_value='<div class="notice notice-warning is-dismissable "><p>'.$_GET['action'].' post '.$mypostid.'</p></div>';
+					}
+
 				}
 
 			}
@@ -133,14 +144,29 @@ function gmuw_fs_index_file_table($posts){
 		$return_value.='<tbody>';
 		foreach ($posts as $post) {
 			$return_value.='<tr>';
+			//title/link
 			$return_value.='<td><a href="'.wp_get_attachment_url($post->ID).'" target="_blank">'.get_the_title($post).'</a></td>';
+			//post ID
 			$return_value.='<td>'. $post->ID.'</td>';
+			//mime type
 			$return_value.='<td>'. $post->post_mime_type.'</td>';
-			$return_value.='<td>'. get_user_by('id', $post->post_author)->user_login .'</td>';
-			$return_value.='<td>'.get_the_modified_date('Y-m-d', $post).'</td>';
+			//user
 			$return_value.='<td>';
-			$return_value.='<a class="button button-primary" href="admin.php?page=gmuw_fs_file_index&action=edit&mypostid='.$post->ID.'">edit</a> ';
-			$return_value.='<a class="button" href="admin.php?page=gmuw_fs_file_index&action=delete&mypostid='.$post->ID.'" onclick="return confirm(\'Are you sure you want to delete this file?\')">delete</a> ';
+			$return_value.= $post->post_author == get_current_user_id() ? '<span class="gmuw_fs_highlight">' : '';
+			$return_value.=get_user_by('id', $post->post_author)->user_login;
+			$return_value.= $post->post_author == get_current_user_id() ? '</span>' : '';
+			$return_value.='</td>';
+			//date modified
+			$return_value.='<td>'.get_the_modified_date('Y-m-d', $post).'</td>';
+			//actions
+			$return_value.='<td>';
+			//does this file belong to the current user?
+			if ($post->post_author == get_current_user_id() || current_user_can('manage_options') ) {
+				//edit button
+				$return_value.='<a class="button button-primary" href="admin.php?page=gmuw_fs_file_index&action=edit&mypostid='.$post->ID.'">edit</a> ';
+				//delete button
+				$return_value.='<a class="button" href="admin.php?page=gmuw_fs_file_index&action=delete&mypostid='.$post->ID.'" onclick="return confirm(\'Are you sure you want to delete this file?\')">delete</a> ';
+			}
 			$return_value.='</td>';
 			$return_value.='</tr>';
 		}
